inotify-sub: Add --monitor, --event, and --historic
The first to move the API closer to inotifywait's example.

--historic lets us support both "I don't care if it already happend"
(set --historic) and "I only want to hear about new events" (don't set
--historic), both of which are useful cases for non-monitor blocking.

Exiting in the non-monitor case is a bit awkward, since we're noticing
that we're done from a subshell (and for historic entries, from a
subsubshell).  This is probably tiptoeing around "too complicated for
a POSIX shell script", but the kill_inotifywait hack gets the job done
for now.

Signed-off-by: W. Trevor King <wking@tremily.us>